requirements.txt
pycrypto
pycryptodome
base58check
ecdsa
uvicorn
pydantic
fastapi